+++ plugins/age/src/index.js
@@ -1,6 +1,6 @@
 export default {
   age: function (unit = 'years') {
-    let now = this.set()
+    let now = this._from()
     let diff = this.diff(now, unit)
     return diff
   }