+++ scratch.js
@@ -2,17 +2,11 @@ import spacetime from './src/01-one/index.js'
 // import { DateTime } from "luxon";
 // import old from './old/src/index.js'
 
-// let a = spacetime('March 28, 1999 20:42:00')
-// a = a.add(45, 'months')
-// console.log(a.iso())
 
-// let s = spacetime('2023-03-12T01:39:00-08:00', 'America/Los_Angeles')
-// console.log(s.time())
+// spacetime.world.methods.now = () => 1554092400000
+// let s = spacetime({ month: 1, offset: 4 })
+// console.log(s.iso())
 
-// console.log(spacetime('foo', 'Asia/Riyadh'))
 
-// let config = { tryLocalTimezone: true, fallbackTz: 'Etc/GMT+9.5' }
-// spacetime.plugin({ config })
-// let s = spacetime('2,000 BC')
-let s = spacetime([2022, 2])
-console.log(s.json())
+let d = spacetime('January 12, 2016')
+console.log(d.daysInMonth(), 31, 'daysInMonth - Jan. 2016')