+++ src/01-one/methods/compute/getEpoch/walk.js
@@ -9,7 +9,7 @@ const ensureEqual = function (a, b) {
     let unit = units[i]
     if (a[unit] !== b[unit]) {
       console.error('\n----\nMis-matched unit in walk:', unit) // eslint-disable-line
-      console.error(a, '\n', b) // eslint-disable-line
+      // console.error(a, '\n', b) // eslint-disable-line
       return false
     }
   }
@@ -23,8 +23,7 @@ const diffDays = function (from, to, world) {
   let diff = 0
   // increment months
   for (let n = from.month; n < to.month; n += 1) {
-    // console.log(`+${months[n - 1].len} for ${months[n - 1].longForm}`)
-    diff += months[n - 1].len
+    diff += months[n].len
     if (n === 2 && isLeapYear(from.year)) {
       diff += 1 //add another
     }