+++ src/01-one/methods/parse/text/units/month.js
@@ -3,11 +3,12 @@ import { isNum } from './_lib.js'
 
 let mapping = {}
 model.months.forEach((_, i) => {
-  mapping[model.months[i].shortForm.toLowerCase()] = i + 1
-  mapping[model.months[i].longForm.toLowerCase()] = i + 1
+  if (model.months[i]) {
+    mapping[model.months[i].shortForm.toLowerCase()] = i
+    mapping[model.months[i].longForm.toLowerCase()] = i
+  }
 })
 mapping.sept = 9//extra
-
 const parseMonth = function (input) {
   if (isNum(input)) {
     return input